+++ modules/importer/oggdude/importers/gear.js
@@ -34,10 +34,11 @@ export default class Gear {
                 tags: [
                   "gear",
                 ],
+                sources: ImportHelpers.getSourcesAsArray(item?.Sources ?? item?.Source),
               },
             };
 
-            data.data.description += ImportHelpers.getSources(item?.Sources ?? item?.Source);
+            //data.data.description += ImportHelpers.getSources(item?.Sources ?? item?.Source);
             const mods = await ImportHelpers.processMods(item);
             if (mods) {
               if (mods.baseMods) {